block_session: add constraining view type
This commit adds the facilities to further constrain the access of a
client to the underlying Block session. It allows for augmenting the
session arguments with an offset in the covered LBA range and in
doing so to reference a block device partial.

This change is going to help implementing a Block session multiplexer
that solely mediates access to a given portion of the device.

Issue #5479.