sculpt: distinguish touch from click for popup
This patch separates the handling of touch input from pointer input in
the logic of closing the popup dialog. For the detection of a touch
outside the popup dialog, nitpicker's new "touch" report is observed
whereas the detection of a click outside the popup is based on
nitpicker's hover report.

Issue #5496
Issue #5485