ld: add dynamic list for stdcxx::delete
When deleting destructors are called for base-library classes which are
part of ld.lib.so, a dummy delete operator from 'new_delete.cc' will be
called because ld.lib.so is compiled with -Bsymbolic-functions which leads
to compile time resolution to the weak dummy functions in cxx. These
functions should resolve to the stdcxx library. Therefore, this commit
adds a dynamic list for the stdcxx::delete operators to ld.lib.so, which
generates a jump-slot relocation, and thus, the linker will try to
resolve the operators at runtime.

issue #5560