libc: add rudimentary aio(4) support
The current implementation is sufficient for running fio as well as
the asynchronous I/O backend of vbox6. It does not support asychronous
notification mechanisms for now and thus relies on polling via
'aio_suspend'.

It is implemented in a standalone fashion and apart from now
potentially triggering additional features in contrib code should not
influence (besides the increased size of the Libc::File_descriptor)
existing ports.

Fixes #5634.