xml_node.h: with_sub_node of given index
The new overload takes an index instead of a type name as argument. This
is useful in situations where one sub node is selected for processing
among several siblings. The scanning via 'for_each_sub_node' yields an
index, which can then be used for 'with_sub_node'. This way, values can
be returned from the scope of 'with_sub_node', which is not possible via
'for_each_sub_node'.

Issue #5611