base-foc: fix the release of cap IDs
The wrong application of CAP_ID_MASK rendered the Allocator_avl::free
ineffective, such that IDs were never re-used. Even though IDs are cheap,
there are not free. Over time, the addition of new allocated IDs to the
Allocator_avl slows down the ID allocation. This could be observed in
test/init_loop. The patch fixes the problem by passing the allocated ID
to free as is.

Issue #5676