pci_decode: report write-combineable BARs
Following the PCI spec, BARs with the prefetchable bit set are reported
safe for write-combined access with the "wc" attribute set to "true" by
default. An exception from this rule are recent Intel graphics devices,
where at least the MMIO registers in BAR 0 must be accessed uncached.
Thus, we report wc as false in BAR 0 for Intel graphics devices.

Issue #5703