window_layouter: rename 'key' attribute to 'name'
The 'press' and 'release' nodes in the configuration of key sequences
use to refer to keys via the 'key' attribute. This patch renames the
attribute to 'name' to be consistent with nitpicker's 'global-key' nodes
and to leverage the special role of the 'name' attribute in HRD syntax.

Compatibility to the original 'key' attribute is preserved.