hw: distinguish assert macro, kernel vs. bootstrap
To aid debugging the kernel, this patch adds backtrace functionality
to assertions of the kernel. The backtrace needs knowledge of the
per-cpu stack and its limitations, which differs between kernel,
core, and bootstrap. Therefore, split assert macro definition
between them, and add the backtrace in the kernel part.

To clearly identify the macro character of the assertion, use
uppercase letters, like elsewhere in the code-base.

Ref #5719
Ref #5725