hw: add assertion to Cpu_context::_deactivate
The removal of the current, active Cpu_context from the scheduler
of a different CPU than the executing one, is a violation of the
integrity of the kernel's execution model. To better detect this
in advance, when extending functionality or otherwise change the
kernel code, add a corresponding assertion to Cpu_context::_deactivate.
The existing assertion of ~Cpu_context can be removed thereby,
because its a special case of the new assertion.

Ref #5719
Ref #5725