depot_deploy: allow for the masking of start nodes
This commit introduces a 'cond_fn' argument to gen_start_nodes, which is
called for each child name and returns the decision whether to generate
the child's start node or not. This allows caller to selectively mask
the presence of each start node in the runtime init configuration.
In particular, the depot-autopilot can use this condition to execute
tests sequentially.

The patch also adds a way for determining the actual children that have
incomplete pkgs rather than only the scarce information that such a
case exists.

Issue #5745