base: avoid invalid cap reasoning for stacks
In core the threads don't have a valid stack dataspace capability, but this
is used to decide whether an entry in the internal stack array is unused.
Beside theoretical potential corruption (when adding a further stack to a
core thread stack), using the base address, which can't be valid at
virtual zero address, enables the detaching of stacks in core thread on
destruction.

Issue #5668