base-nova: introduce Pd_selector in vm session
This commit encapsulates the pd selector, and its insertion, removal
in an utility struct called: Core::Vm_session_component::Pd_selector.
Thereby the removal of the underlying pd can be synchronized with
the destruction of the Core::Guest_memory utility, which is book-keeping
all attachments. Otherwise, the pd gets removed before the guest memory
attachments can be cleaned up accordingly.

This commit is a necessary follow-up work of issue #5665.